fix: i16 and i32 string form was detected as i8

+++ tot_spec/src/codegen/fixtures/rs_serde/keyword.rs
@@ -4,5 +4,5 @@ pub struct TestRustKeyword {
     #[serde(rename = "fn")]
     pub fn_: std::option::Option<std::string::String>,
     #[serde(rename = "const")]
-    pub const_: std::option::Option<i8>,
+    pub const_: std::option::Option<i32>,
 }
